+++ module-menu/src/test/java/kitchenpos/ApplicationTest.java
@@ -1,0 +1,11 @@
+package kitchenpos;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApplicationTest {
+    @Test
+    void contextLoads() {
+    }
+}
