+++ module-menu/src/test/java/kitchenpos/menu/service/MenuGroupServiceTest.java
@@ -6,9 +6,8 @@ import static org.assertj.core.api.SoftAssertions.assertSoftly;
 import java.util.List;
 import kitchenpos.menu.service.dto.MenuGroupRequest;
 import kitchenpos.menu.service.dto.MenuGroupResponse;
-import kitchenpos.menu.service.MenuGroupService;
-import kitchenpos.supports.MenuGroupFixture;
 import kitchenpos.supports.IntegrationTest;
+import kitchenpos.MenuGroupFixture;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
