+++ module-menu/src/test/java/kitchenpos/menu/service/MenuServiceTest.java
@@ -6,19 +6,17 @@ import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
 import java.math.BigDecimal;
 import java.util.List;
+import kitchenpos.common.vo.Price;
+import kitchenpos.menu.service.dto.MenuGroupResponse;
 import kitchenpos.menu.service.dto.MenuProductRequest;
 import kitchenpos.menu.service.dto.MenuRequest;
-import kitchenpos.menu.service.dto.MenuGroupResponse;
 import kitchenpos.menu.service.dto.MenuResponse;
-import kitchenpos.product.service.dto.ProductResponse;
-import kitchenpos.common.vo.Price;
-import kitchenpos.menu.service.MenuService;
-import kitchenpos.menu.service.MenuGroupService;
 import kitchenpos.product.service.ProductService;
+import kitchenpos.product.service.dto.ProductResponse;
 import kitchenpos.supports.IntegrationTest;
-import kitchenpos.supports.MenuFixture;
-import kitchenpos.supports.MenuGroupFixture;
-import kitchenpos.supports.ProductFixture;
+import kitchenpos.MenuFixture;
+import kitchenpos.MenuGroupFixture;
+import kitchenpos.ProductFixture;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
