+++ module-menu/src/test/resources/truncate.sql
@@ -1,0 +1,6 @@
+SET REFERENTIAL_INTEGRITY FALSE;
+TRUNCATE TABLE menu RESTART IDENTITY;
+TRUNCATE TABLE menu_group RESTART IDENTITY;
+TRUNCATE TABLE menu_product RESTART IDENTITY;
+TRUNCATE TABLE product RESTART IDENTITY;
+SET REFERENTIAL_INTEGRITY TRUE;
