+++ module-menu/src/testFixtures/java/kitchenpos/MenuFixture.java
@@ -1,12 +1,12 @@
-package kitchenpos.supports;
+package kitchenpos;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
+import kitchenpos.common.vo.Price;
 import kitchenpos.menu.service.dto.MenuProductRequest;
 import kitchenpos.menu.service.dto.MenuRequest;
 import kitchenpos.product.service.dto.ProductResponse;
-import kitchenpos.common.vo.Price;
 
 public class MenuFixture {
