+++ module-order/src/test/java/kitchenpos/order/service/OrderServiceTest.java
@@ -6,6 +6,9 @@ import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
 import java.util.Collections;
 import java.util.List;
+import kitchenpos.MenuFixture;
+import kitchenpos.ProductFixture;
+import kitchenpos.MenuGroupFixture;
 import kitchenpos.menu.service.MenuGroupService;
 import kitchenpos.menu.service.MenuService;
 import kitchenpos.menu.service.dto.MenuGroupResponse;
@@ -13,8 +16,6 @@ import kitchenpos.menu.service.dto.MenuRequest;
 import kitchenpos.menu.service.dto.MenuResponse;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.repository.OrderRepository;
-import kitchenpos.order.service.OrderService;
-import kitchenpos.order.service.TableService;
 import kitchenpos.order.service.dto.OrderRequest;
 import kitchenpos.order.service.dto.OrderResponse;
 import kitchenpos.order.service.dto.OrderStatusRequest;
@@ -22,11 +23,8 @@ import kitchenpos.order.service.dto.TableResponse;
 import kitchenpos.product.service.ProductService;
 import kitchenpos.product.service.dto.ProductResponse;
 import kitchenpos.supports.IntegrationTest;
-import kitchenpos.supports.MenuFixture;
-import kitchenpos.supports.MenuGroupFixture;
-import kitchenpos.supports.OrderFixture;
-import kitchenpos.supports.OrderTableFixture;
-import kitchenpos.supports.ProductFixture;
+import kitchenpos.OrderFixture;
+import kitchenpos.OrderTableFixture;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
