+++ module-order/src/test/java/kitchenpos/order/service/TableServiceTest.java
@@ -5,27 +5,24 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
 import java.util.List;
-import kitchenpos.order.service.dto.OrderStatusRequest;
-import kitchenpos.order.service.dto.TableGroupRequest;
-import kitchenpos.order.service.dto.TableRequest;
+import kitchenpos.MenuFixture;
+import kitchenpos.MenuGroupFixture;
+import kitchenpos.ProductFixture;
+import kitchenpos.menu.service.MenuGroupService;
+import kitchenpos.menu.service.MenuService;
 import kitchenpos.menu.service.dto.MenuGroupResponse;
 import kitchenpos.menu.service.dto.MenuResponse;
 import kitchenpos.order.service.dto.OrderResponse;
-import kitchenpos.product.service.dto.ProductResponse;
+import kitchenpos.order.service.dto.OrderStatusRequest;
+import kitchenpos.order.service.dto.TableGroupRequest;
+import kitchenpos.order.service.dto.TableRequest;
 import kitchenpos.order.service.dto.TableResponse;
-import kitchenpos.menu.service.MenuService;
-import kitchenpos.menu.service.MenuGroupService;
-import kitchenpos.order.service.OrderService;
 import kitchenpos.product.service.ProductService;
+import kitchenpos.product.service.dto.ProductResponse;
 import kitchenpos.supports.IntegrationTest;
-import kitchenpos.supports.MenuFixture;
-import kitchenpos.supports.MenuGroupFixture;
-import kitchenpos.supports.OrderFixture;
-import kitchenpos.supports.OrderTableFixture;
-import kitchenpos.supports.ProductFixture;
-import kitchenpos.supports.TableGroupFixture;
-import kitchenpos.order.service.TableService;
-import kitchenpos.order.service.TableGroupService;
+import kitchenpos.OrderFixture;
+import kitchenpos.OrderTableFixture;
+import kitchenpos.TableGroupFixture;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
